+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - feat: add `backgroundColor` to `NesContainer`
 - fix: `NesController` throwing initialization errors
+- feat: add `NesIcons.openFolder`
+- feat: add `NesIcons.closedFolder`
 
 # 0.2.0